+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## [NEXT_VERSION] - [UNRELEASED]
-* BUG: Example next bug fixed.
+* ENH: New user created email template.
+* ENH: Admin created user set password.
+* ENH: Email Templates - Clickable variables.
+* ENH: Email Templates - Add button to send a test email.
 
 ## [1.1.0] - 2026-02-08
 * FEA: Add send new email modal for super admins.